Task 3, fixed bug

+++ src/main/java/dao/UserDAO.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserDAO {
     List<User> getAllUsersDAO();
     void addUserDAO(User user);
-    void deleteUserDAO(int id);
+    void deleteUserDAO(Long id);
     void updateUserDAO(User newUser);
-    User getUserById(int id);
+    User getUserById(Long id);
 }
