+++ src/main/java/service/UserService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface UserService {
     void addUser(User user);
-    User getUserById(int id);
-    void deleteUser(int id);
+    User getUserById(Long id);
+    void deleteUser(Long id);
     void updateUser(User user);
     List<User> getAllUsers();
 }
